Milestone 1: the kernel builds into a wasm module
west build -b wasm_node samples/hello_world now produces a 300 KB WebAssembly
binary module with the intended interface: six zephyr_host imports and the
boot, switch-block, pending-IRQ and thread-entry exports, plus __stack_pointer
as a mutable global.

The kernel, the minimal libc and both drivers compile for wasm32 essentially
untouched. The work was at the edges.

The offsets generator needed the Kconfig header force-included, the target
triple passed, an explicit path to gen_offset.h since Zephyr builds that
include from the arch root, and ordering after the syscall headers because it
compiles its source during header generation rather than reading a later
object.

GEN_ABSOLUTE_SYM_KCONFIG cannot use the data form: its callers pass names that
are themselves macros, which only works when the name is stringified before
expansion. Nothing reads those symbols at run time, so it is a no-op here.

There are four arch dispatch headers rather than two; patch 0002 now covers
three, exception.h being the one missed.

+++ drivers/timer/wasm_host_timer.c
@@ -39,11 +39,11 @@ static void timer_isr(const void *arg)
 	}
 }
 
-void sys_clock_set_timeout(int32_t ticks, bool idle)
+void sys_clock_set_timeout(uint32_t ticks, bool idle)
 {
 	ARG_UNUSED(idle);
 
-	if (ticks == K_TICKS_FOREVER) {
+	if (ticks == (uint32_t)K_TICKS_FOREVER) {
 		/* Nothing to wake for. Leaving no alarm set is what lets the
 		 * host decide the run is over.
 		 */
@@ -51,8 +51,8 @@ void sys_clock_set_timeout(int32_t ticks, bool idle)
 		return;
 	}
 
-	if (ticks < 1) {
-		ticks = 1;
+	if (ticks < 1U) {
+		ticks = 1U;
 	}
 
 	wasm_host_set_alarm_ns(last_announced_ns + (int64_t)ticks * NSEC_PER_TICK);
